FXML fix, add Setup for How to video
